feat: normalize expense meta into typed detail tables
Replace the untyped Expense.meta Json column with typed 1:1 relations
(travel_expense_detail, food_expense_detail). The migration backfills
existing rows idempotently; meta is kept nullable and deprecated until
a follow-up contract migration drops it (expand/contract, zero loss).

This removes the three divergent meta shape definitions (lib/validators,
expense.validators, apps/api generate.ts — the last silently dropped
food deduction fields), the silent-fallback merge logic in the expense
update path, and the CSV export bug that rendered meta as
'[object Object]'. Audit diffs now record travelDetail/foodDetail keys
instead of meta.

Deletes apps/web/src/server/pdf/summary.ts: it was an unreferenced
duplicate of the apps/api PDF path and depended on the removed meta
schemas.

+++ apps/web/src/server/modules/expense/expense.validators.ts
@@ -48,7 +48,7 @@ export const createTravelExpenseSchema = baseCreateExpenseSchema.and(
 
 export const createFoodExpenseSchema = baseCreateExpenseSchema.and(
 	z.object({
-		days: z.number().min(1),
+		days: z.number().int().min(1),
 		breakfastDeduction: z.number().min(0).multipleOf(0.01),
 		lunchDeduction: z.number().min(0).multipleOf(0.01),
 		dinnerDeduction: z.number().min(0).multipleOf(0.01),
@@ -63,7 +63,7 @@ export const updateExpenseSchema = z.object({
 	from: z.string().min(1).optional(),
 	to: z.string().min(1).optional(),
 	distance: z.number().min(1).optional(),
-	days: z.number().min(1).optional(),
+	days: z.number().int().min(1).optional(),
 	breakfastDeduction: z.number().min(0).multipleOf(0.01).optional(),
 	lunchDeduction: z.number().min(0).multipleOf(0.01).optional(),
 	dinnerDeduction: z.number().min(0).multipleOf(0.01).optional(),
